passing IntegerCacheTest after dirty-fix of sun.misc.VM with class file
related to wwu-pi/muggl#40
Initialization of the IntegerCache calls sun.misc.VM.getSavedProperty(String key)
which in the original rt.jar does a check if systems' properties are empty (which they are in muggl at this time of
writing). It will then emit a IllegalStateException("Should be non-empty if initialized").

To cirumvent this as a dirty fix, there is a patched VM.java in muggl-core/sun/misc that does not do this check
and hence these checks here should pass.